Move main command to 'build' module
For now, this simply lets us run the build command in the same way we
run the check command, namely:

    $ python -m towncrier.build

However, it will let us do far more powerful stuff shortly.
